Move Usage into its own module.
